requirements.txt
absl_py==1.0.0+computecanada
anyio==4.4.0+computecanada
argon2_cffi==23.1.0+computecanada
argon2_cffi_bindings==21.2.0+computecanada
arrow==1.3.0+computecanada
asttokens==2.4.1+computecanada
astunparse==1.6.3+computecanada
async_lru==2.0.4+computecanada
attrs==24.2.0+computecanada
babel==2.16.0+computecanada
backcall==0.2.0+computecanada
beautifulsoup4==4.12.3+computecanada
bleach==6.1.0+computecanada
cachetools==5.2.0+computecanada
certifi==2022.5.18.1+computecanada
cffi
charset_normalizer==2.0.12+computecanada
cloudpickle==2.1.0+computecanada
comm==0.2.2+computecanada
contourpy==1.3.0+computecanada
cycler==0.11.0+computecanada
debugpy==1.8.5+computecanada
decorator==5.1.1+computecanada
defusedxml==0.7.1+computecanada
exceptiongroup==1.2.2+computecanada
executing==2.0.1+computecanada
fastjsonschema==2.20.0+computecanada
flatbuffers==24.3.25+computecanada
fonttools
fqdn==1.5.1+computecanada
gast==0.4.0+computecanada
gin_config==0.5.0+computecanada
google-pasta==0.2.0+computecanada
google_auth==2.35.0+computecanada
google_auth_oauthlib==1.2.1+computecanada
grpcio==1.67.0+computecanada
gym==0.24.0+computecanada
gym-minigrid==1.0.1+computecanada
gym_notices==0.0.8+computecanada
h11==0.14.0+computecanada
h5py
httpcore==1.0.5+computecanada
httpx==0.27.2+computecanada
idna==3.3+computecanada
importlib_metadata==7.2.1+computecanada
importlib_resources==6.4.5+computecanada
ipykernel==6.29.5+computecanada
ipython==8.22.2+computecanada
ipython_genutils==0.2.0+computecanada
ipywidgets==8.1.5+computecanada
isoduration==20.11.0+computecanada
jedi==0.19.1+computecanada
jinja2==3.1.4+computecanada
json5==0.9.25+computecanada
jsonpointer==3.0.0+computecanada
jsonschema==4.23.0+computecanada
jsonschema_specifications==2023.12.1+computecanada
jupyter==1.0.0+computecanada
jupyter-console==6.4.0+computecanada
jupyter_client==8.6.2+computecanada
jupyter_core==5.7.2+computecanada
jupyter_events==0.10.0+computecanada
jupyter_lsp==2.2.5+computecanada
jupyter_server==2.14.2+computecanada
jupyter_server_terminals==0.5.3+computecanada
jupyterlab==4.1.3+computecanada
jupyterlab_pygments==0.3.0+computecanada
jupyterlab_server==2.27.3+computecanada
jupyterlab_widgets==3.0.13+computecanada
keras
Keras-Preprocessing==1.1.2+computecanada
kiwisolver==1.4.7+computecanada
libclang==14.0.1+computecanada
Markdown==3.3.7+computecanada
MarkupSafe
matplotlib==3.9.2+computecanada
matplotlib_inline==0.1.7+computecanada
mistune==3.0.2+computecanada
ml_dtypes
nbclient==0.10.0+computecanada
nbconvert==7.16.4+computecanada
nbformat==5.10.4+computecanada
nest_asyncio==1.6.0+computecanada
notebook==7.1.1+computecanada
notebook_shim==0.2.4+computecanada
numpy
oauthlib==3.2.0+computecanada
opt_einsum==3.3.0+computecanada
overrides==7.7.0+computecanada
packaging==24.1+computecanada
pandocfilters==1.5.1+computecanada
parso==0.8.4+computecanada
pexpect==4.9.0+computecanada
pickleshare==0.7.5+computecanada
Pillow
pkgutil_resolve_name==1.3.10+computecanada
platformdirs==4.2.2+computecanada
prometheus_client==0.20.0+computecanada
prompt_toolkit==3.0.47+computecanada
protobuf==4.25.4+computecanada
psutil==5.9.4+computecanada
ptyprocess==0.7.0+computecanada
pure_eval==0.2.3+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pycparser==2.22+computecanada
pygments==2.18.0+computecanada
pyparsing==3.0.9+computecanada
python-dateutil==2.8.2+computecanada
python_json_logger==2.0.7+computecanada
pytz==2024.1+computecanada
PyYAML==6.0.1+computecanada
pyzmq
qtconsole==5.3.2+computecanada
QtPy==2.2.0+computecanada
referencing==0.35.1+computecanada
requests==2.32.3+computecanada
requests_oauthlib==1.3.1+computecanada
rfc3339_validator==0.1.4+computecanada
rfc3986_validator==0.1.1+computecanada
rpds_py
rsa==4.8+computecanada
scipy
Send2Trash==1.8.3+computecanada
six==1.16.0+computecanada
sniffio==1.3.1+computecanada
soupsieve==2.6+computecanada
stack_data==0.6.3+computecanada
tensorboard
tensorboard_data_server==0.7.2+computecanada
tensorboard_plugin_wit==1.8.1+computecanada
tensorflow
tensorflow_estimator==2.15.0+computecanada
termcolor==1.1.0+computecanada
terminado==0.18.1+computecanada
tinycss2==1.3.0+computecanada
tomli==2.0.1+computecanada
tornado==6.3.3+computecanada
traitlets==5.14.3+computecanada
types_python_dateutil==2.9.0.20241003+computecanada
typing_extensions==4.2.0+computecanada
uri_template==1.3.0+computecanada
urllib3==1.26.9+computecanada
wcwidth==0.2.13+computecanada
webcolors==24.11.1+computecanada
webencodings==0.5.1+computecanada
websocket_client==1.8.0+computecanada
Werkzeug==2.1.2+computecanada
widgetsnbextension==4.0.13+computecanada
wrapt
zipp==3.8.0+computecanada
